LEDMatrix is now a segment of a larger LEDStrip
LEDMatrix can now be put alongside other LEDStripSegments instead of creating its own AddressibleLED object.

Other refactoring and updated examples.

Boolean parameter clearOthers was removed because it was unnecessary and meant a lot of extra code.  Just use clear() or off() before the method if clearOthers was set to true.